% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapMetToGSMN.R
\name{annotate_chemical_properties}
\alias{annotate_chemical_properties}
\title{Function to annotate InputData with chemical properties.}
\usage{
annotate_chemical_properties(inputData)
}
\arguments{
\item{InputData}{QFeatures experiment}
}
\value{
QFeatures experiment similar to input and with chemical properties added to peakList RowData
}
\description{
This function retrieves chemical properties such as SMILES, InChI and InChIKey
from ChEBI using the package webchem. Notice that there are ChEBI entries which have SMILES but
no InChI/InChIKey annotated because of the possibility to represent variable ends with an *
in SMILES but not in InChI/InChIKey.
}
\author{
Sarah Scharfenberg
}
