% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkComparisonFunctions.R
\name{basicNetStats}
\alias{basicNetStats}
\title{Calculate basic network stats}
\usage{
basicNetStats(network, netName)
}
\description{
Calculate basic network stats
}
\author{
Elva Novoa, \email{elva-maria.novoa-del-toro@inrae.fr}
}
