% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiLayerFunctions.R
\name{buildCorrNet}
\alias{buildCorrNet}
\title{Build the correlation network}
\usage{
buildCorrNet(inputData, directed, corrModel, corrThresh)
}
\description{
Build the correlation network
}
\author{
Elva Novoa, \email{elva-maria.novoa-del-toro@inrae.fr}
}
