% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiLayerFunctions.R
\name{buildExpNet}
\alias{buildExpNet}
\title{Build experimental networks}
\usage{
buildExpNet(
  inputData,
  net2Build = "all",
  directed = FALSE,
  ppmMass = 10,
  ppmSpec = 0,
  tol = 0.005,
  corrModel = "pearson",
  corrThresh = 0.25
)
}
\arguments{
\item{inputData}{\code{list}, list returned by the \code{loadInputData} function}

\item{net2Build}{\code{list}, list of experimental networks to build, according to the following
options:
"all" - generates 3 experimental networks: mass difference, correlation,
and spectral similarity,
"m" - builds the mass difference network,
"s" - builds the spectral similarity network,
"c" - builds the correlation network.
It is to note that combinations of m, s, and c are possible, writing them
separated by commas, e.g., c("m", "c") would generate the mass
difference, and correlation networks. The default value is "all"}

\item{directed}{\code{boolean}, boolean value (TRUE/FALSE). If TRUE then the networks that are
generated will be directed, and undirected otherwise. The default value is
FALSE (i.e., undirected network)}

\item{ppmMass}{\code{numeric}, (optional) allowed error for mass differences calculus. It is
only needed if the mass difference network is to be built. The default value
is 10}

\item{ppmSpec}{\code{numeric}, (optional) relative allowed error for spectral similarity
calculus. It is only needed if the spectral similarity network is to be
built. The default value is 0}

\item{tol}{\code{numeric}, (optional) absolute tolerance for spectral similarity calculus.
It is only needed if the spectral similarity network is to be built. The
default value is 0.005}

\item{corrModel}{\code{vector}, (optional) character vector containing the model(s) to be used for
the correlation calculus in MetNet. Please check the available models in the
current version of MetNet. There are 10 models available in version 1.14.0:
"lasso", "randomForest", "clr", "aracne", "pearson", "pearson_partial",
"spearman", "spearman_partial", "ggm", and "bayes". This parameter is only
needed if the correlation network is to be built. The default value is
"pearson"}

\item{corrThresh}{\code{numeric}, (optional) floating point number indicating the correlation
threshold to consider that two features are correlated. It is only needed if
the correlation network is to be built. The default value is 0.25 (i.e., at
least 25\% of correlation between the abundance values, either positive or
negative)}
}
\value{
List of experimental networks as igraph objects
}
\description{
The function \code{buildExpNet} builds experimental networks, based on spectral
similarity, correlation, and mass difference
}
\examples{
# See the MultiLayerNetwork vignette

}
\author{
Elva Novoa, \email{elva-maria.novoa-del-toro@inrae.fr}
}
