% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiLayerFunctions.R
\name{buildMassDiffNet}
\alias{buildMassDiffNet}
\title{Build the mass difference network}
\usage{
buildMassDiffNet(inputData, ppmMass, directed)
}
\description{
Build the mass difference network
}
\author{
Elva Novoa, \email{elva-maria.novoa-del-toro@inrae.fr}
}
