% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiLayerFunctions.R
\name{buildSpecSimNet}
\alias{buildSpecSimNet}
\title{Build the spectral similarity network}
\usage{
buildSpecSimNet(inputData, tol, ppmSpec, massDiff, directed)
}
\description{
Build the spectral similarity network
}
\author{
Elva Novoa, \email{elva-maria.novoa-del-toro@inrae.fr}
}
