% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiLayerFunctions.R
\name{calculateMultiLayerStats}
\alias{calculateMultiLayerStats}
\title{Calculate some statistics of the multi-layer network}
\usage{
calculateMultiLayerStats(multiLayer, inputData)
}
\arguments{
\item{multiLayer}{\code{list}, list containing the multi-layer network as returned by the
\code{makeMultiLayer} function}

\item{inputData}{\code{list}, list returned by the \code{loadInputData} function}
}
\value{
Nothing, but it creates several plots in the resPath directory
}
\description{
Function to calculate some statistics of the multi-layer network
}
\examples{
# See the MultiLayerNetwork vignette

}
\author{
Elva Novoa, \email{elva-maria.novoa-del-toro@inrae.fr}
}
