% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkComparisonFunctions.R
\name{calculateOverlap}
\alias{calculateOverlap}
\title{Calculate, create a table and plot statistics}
\usage{
calculateOverlap(networks)
}
\arguments{
\item{networks}{\code{list}, igraph objects to analyze}
}
\value{
Nothing, but it prints the corresponding plots
}
\description{
The function \code{calculateOverlap} serves to calculate and print plots of the
overlap of nodes and edges
}
\examples{
# See the NetworkComparison vignette

}
\author{
Elva Novoa, \email{elva-maria.novoa-del-toro@inrae.fr}
}
