% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkComparisonFunctions.R
\name{calculNetStats}
\alias{calculNetStats}
\title{Calculate, create a table and plot statistics}
\usage{
calculNetStats(net)
}
\arguments{
\item{net}{\code{list}, igraph objects, as returned by readNet()}
}
\value{
Data frame with the statistics
}
\description{
The function \code{calculNetStats} serves to calculate, create a table and plot
the following statistics:
\itemize{
\item Density (no self loops are considered)
\item Diameter (if there is more than one connected component, the largest
diameter will be returned, independently of the size of the connected
component)
\item Average degree
\item Average path length
\item Clustering coefficient (the networks are considered as undirected)
It also plots the degree distribution, and upset plots of the overlap of
nodes and edges reads all the networks stored in a given directory
}
}
\examples{
# See the NetworkComparison vignette

}
\author{
Elva Novoa, \email{elva-maria.novoa-del-toro@inrae.fr}
}
