% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiLayerFunctions.R
\name{cytoscapeVis}
\alias{cytoscapeVis}
\title{Visualize in Cytoscape}
\usage{
cytoscapeVis(nodes, edges, resPath, mainType = "GSMN", fixedColors = TRUE)
}
\description{
Visualize in Cytoscape
}
\author{
Elva Novoa, \email{elva-maria.novoa-del-toro@inrae.fr}
}
