% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapMetToGSMN.R
\name{.mapMetToGSMN_inchikey}
\alias{.mapMetToGSMN_inchikey}
\title{Map features to metabolites if the InChIKey is the same.}
\usage{
.mapMetToGSMN_inchikey(inputData, gsmnMetaFile = NULL)
}
\arguments{
\item{inputData}{\code{list}, list returned by the \code{loadInputData} function}
}
\description{
Function to map the identified experimental nodes to the
corresponding GSMN nodes, by InChIKey comparison.
}
\author{
Sarah Scharfenberg
}
