% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export2gml.R
\name{export2gml}
\alias{export2gml}
\alias{exportNet2gml}
\title{Create a network and export to gml from adjacency list}
\usage{
exportNet2gml(x, file, select = F, ...)
}
\arguments{
\item{x}{\code{data.frame} adjacency list having \code{Var1} and \code{Var2} in the first
two columns and additional (edge) attributes in following columns.}

\item{select}{\code{character}, the default is \code{FALSE} and returns a gml having edge
attributes of all availabe columns of \code{x}. A \code{character} value or
string may be used to select one ore more columns by their columnnames
that should be saved as edge attributes.}

\item{name}{\code{character}, the name which the gml should have, e.g. "mz"
produces the file 'mz.gml' in the current working directory.}
}
\value{
\code{.gml} file saved in current working directory with selected name
using \code{file }.
}
\description{
The function \code{export2gml} creates a network from adjacency lists.
The gml is saved in the current working directory having a
selected name (\code{name}). All or selected edge attributes may be
selected using the parameter \code{select}.
}
\details{
´\code{export2gml} uses adjacency matrixes to create first a network
and then saving it as gml in the current working directory.
The name of the gml file is selected by the User with the
\code{name} parameter.
The networks contains the feature ID's as nodes. The edges
contains corresponding source  feature ID (sourceID) and target
feature ID (targetID) and either all column attributes as additional
edge attributes or selected ones if \code{select} was used.
}
\examples{
 ######  example to be addded! 


}
\author{
Liesa Salzer, \email{liesa.salzer@helmholtz-muenchen.de}
}
