% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkComparisonFunctions.R
\name{getBetweenness}
\alias{getBetweenness}
\title{Calculate the betweenness of the nodes in a network}
\usage{
getBetweenness(network, netName)
}
\description{
Calculate the betweenness of the nodes in a network
}
\author{
Elva Novoa, \email{elva-maria.novoa-del-toro@inrae.fr}
}
