% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapMetToGSMN.R
\name{getChemicalPropertiesFromChebi}
\alias{getChemicalPropertiesFromChebi}
\title{Function to retrieve chemical properties such as InChIKey, Smiles
from a GSMN file containing ChEBI Ids.}
\usage{
getChemicalPropertiesFromChebi(fileIn, fileOut = NULL)
}
\description{
Function to retrieve chemical properties such as InChIKey, Smiles
from a GSMN file containing ChEBI Ids.
}
