% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkComparisonFunctions.R
\name{getClosenessCC}
\alias{getClosenessCC}
\title{Calculate the closeness of each connected component of a network}
\usage{
getClosenessCC(cc, network, netName)
}
\description{
Calculate the closeness of each connected component of a network
}
\author{
Elva Novoa, \email{elva-maria.novoa-del-toro@inrae.fr}
}
