% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiLayerFunctions.R
\name{getColorTable}
\alias{getColorTable}
\title{Get the table of colors for the nodes and edges}
\usage{
getColorTable(nodeTypes, edgeTypes, fixedColors)
}
\description{
Get the table of colors for the nodes and edges
}
\author{
Elva Novoa, \email{elva-maria.novoa-del-toro@inrae.fr}
}
