% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiLayerFunctions.R
\name{getEdgeList}
\alias{getEdgeList}
\title{Get the list of edges in a multi-layer network}
\usage{
getEdgeList(multiLayer)
}
\description{
Get the list of edges in a multi-layer network
}
\author{
Elva Novoa, \email{elva-maria.novoa-del-toro@inrae.fr}
}
