% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkComparisonFunctions.R
\name{getNodeDeg}
\alias{getNodeDeg}
\title{Get the nodes' degrees}
\usage{
getNodeDeg(net)
}
\description{
Get the nodes' degrees
}
\author{
Elva Novoa, \email{elva-maria.novoa-del-toro@inrae.fr}
}
