% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiLayerFunctions.R
\name{getNodeList}
\alias{getNodeList}
\title{Get the list of nodes in a multi-layer network}
\usage{
getNodeList(multiLayer)
}
\description{
Get the list of nodes in a multi-layer network
}
\author{
Elva Novoa, \email{elva-maria.novoa-del-toro@inrae.fr}
}
