% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkComparisonFunctions.R
\name{getOverlappingNodes}
\alias{getOverlappingNodes}
\title{Get the overlapping nodes}
\usage{
getOverlappingNodes(networks, networksIndex)
}
\arguments{
\item{networks}{\code{list}, list of networks (igraph objects)}

\item{networksIndex}{\code{list}, list of networks' index to consider for the overlap, e.g.,
c(1, 3) to take the first and third networks}
}
\value{
List of overlapping nodes
}
\description{
The function \code{getOverlappingNodes} gets the overlapping nodes from a given
set of input networks
}
\examples{
# See the NetworkComparison vignette

}
\author{
Elva Novoa, \email{elva-maria.novoa-del-toro@inrae.fr}
}
