% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkComparisonFunctions.R
\name{getPlotLabels}
\alias{getPlotLabels}
\title{Get the labels to plot in the box plot}
\usage{
getPlotLabels(data, stat2Plot, bestValue)
}
\description{
Get the labels to plot in the box plot
}
\author{
Elva Novoa, \email{elva-maria.novoa-del-toro@inrae.fr}
}
