% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiLayerFunctions.R
\name{makeBarPlot}
\alias{makeBarPlot}
\title{Make and save a bar plot}
\usage{
makeBarPlot(resPath, data, xAxis, yAxis, title, vertical = TRUE)
}
\description{
Make and save a bar plot
}
\author{
Elva Novoa, \email{elva-maria.novoa-del-toro@inrae.fr}
}
