% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkComparisonFunctions.R
\name{makeBoxPlot}
\alias{makeBoxPlot}
\title{Make and print a boxplot}
\usage{
makeBoxPlot(data, stat2Plot, bestValue = "max", printLab = FALSE, log = FALSE)
}
\description{
Make and print a boxplot
}
\author{
Elva Novoa, \email{elva-maria.novoa-del-toro@inrae.fr}
}
