% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkComparisonFunctions.R
\name{makeHist}
\alias{makeHist}
\title{Calculate and print a set of histograms}
\usage{
makeHist(data, stat2Plot, binWidth = 1, minDeg = 0, maxDeg = 10)
}
\description{
Calculate and print a set of histograms
}
\author{
Elva Novoa, \email{elva-maria.novoa-del-toro@inrae.fr}
}
