% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiLayerFunctions.R
\name{makeMultiLayer}
\alias{makeMultiLayer}
\title{Make a multi-layer network}
\usage{
makeMultiLayer(inputData, expNetworks, pathToMappings)
}
\arguments{
\item{inputData}{\code{list}, list returned by the \code{loadInputData} function}

\item{expNetworks}{\code{list}, list returned by the buildExpNet function}

\item{pathToMappings}{\code{path}, path to folder containing the file(s) with the mapping between the
metabolites from the  experimental networks and those from the GSMN. Such
mapping  can be obtained with tools such as Metabolomics2Networks (see
the MultiLayerNetwork vignette). The file must contain at least 4 columns,
separated by tabs: "metabolite name" (name of the features from the peak
list), "mapped on id" (id of the corresponding metabolite from the GSMN),
"distance" (distance between the feature and its corresponding metabolite.
Note. The distance is equal to zero when the mapping is exact), and "chebi"
(ChEBI id of the corresponding metabolite from the GSMN)}
}
\value{
Multi-layer network in list format containing 3 named elements:
}
\description{
Function to make the multi-layer network by connecting the Genome-Scale
Metabolic Network (GSMN) and the experimental networks
}
\examples{
# See the MultiLayerNetwork vignette

}
\author{
Elva Novoa, \email{elva-maria.novoa-del-toro@inrae.fr}
}
