% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkComparisonFunctions.R
\name{makeScatterPlot}
\alias{makeScatterPlot}
\title{Make and print a scatter plot}
\usage{
makeScatterPlot(netStats, stat2Plot, title = "", printLab = FALSE)
}
\description{
Make and print a scatter plot
}
\author{
Elva Novoa, \email{elva-maria.novoa-del-toro@inrae.fr}
}
