% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkComparisonFunctions.R
\name{makeTableCCSize}
\alias{makeTableCCSize}
\title{Make a table of the sizes of the connected components}
\usage{
makeTableCCSize(cc)
}
\description{
Make a table of the sizes of the connected components
}
\author{
Elva Novoa, \email{elva-maria.novoa-del-toro@inrae.fr}
}
