% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkComparisonFunctions.R
\name{makeToyNet}
\alias{makeToyNet}
\title{Generate toy networks}
\usage{
makeToyNet(netDir)
}
\arguments{
\item{netDir}{\code{Path}, directory to store the toy networks}
}
\value{
Nothing, but it creates four new files in \code{netDir}
}
\description{
The function \code{makeToyNet} generates 4 toy networks.
}
\examples{
# See the NetworkComparison vignette

}
\author{
Elva Novoa, \email{elva-maria.novoa-del-toro@inrae.fr}
}
