% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkComparisonFunctions.R
\name{makeUpsetPlot}
\alias{makeUpsetPlot}
\title{Make and print an upset plot}
\usage{
makeUpsetPlot(
  dataToPlot,
  title = "Overlap",
  xLabel = "Set Size",
  yLabel = "Intersection Size"
)
}
\description{
Make and print an upset plot
}
\author{
Elva Novoa, \email{elva-maria.novoa-del-toro@inrae.fr}
}
