% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiLayerFunctions.R
\name{mapMetToGSMN}
\alias{mapMetToGSMN}
\title{Map the experimental nodes to GSMN nodes}
\usage{
mapMetToGSMN(inputData, resFile = "Res_Met2Net_MappedMet.txt")
}
\arguments{
\item{inputData}{\code{list}, list returned by the \code{loadInputData} function}

\item{resFile}{\code{character}, file name for the resulting mappings file. The default value is
"Res_Met2Net_MappedMet.txt"}
}
\value{
Data frame with the mappings and ontology-based distances.
}
\description{
Function to map the identified experimental nodes to the
corresponding GSMN nodes, using the CHeBI ontology. This function calls the
Metabolomics2Networks Python package
}
\examples{
# See the MultiLayerNetwork vignette

}
\author{
Elva Novoa, \email{elva-maria.novoa-del-toro@inrae.fr}
}
