% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkComparisonFunctions.R
\name{printNothing}
\alias{printNothing}
\title{Print empty lines}
\usage{
printNothing(n)
}
\description{
Print empty lines
}
\author{
Elva Novoa, \email{elva-maria.novoa-del-toro@inrae.fr}
}
