% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkComparisonFunctions.R
\name{printStatsPlots}
\alias{printStatsPlots}
\title{Calculate, create a table and plot statistics}
\usage{
printStatsPlots(stats)
}
\arguments{
\item{stats}{\code{list}, results, as returned by the \code{calculNetStats} function}
}
\value{
Nothing, but it prints several plots
}
\description{
The function \code{printStatsPlots} makes and prints several plots
}
\examples{
# See the NetworkComparison vignette

}
\author{
Elva Novoa, \email{elva-maria.novoa-del-toro@inrae.fr}
}
