% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qfeat.R
\name{qfeat_annotation}
\alias{qfeat_annotation}
\alias{qfeat_annotaion}
\title{Annotation of Qfeatures}
\usage{
qfeat_annotation(x, assay_name = "features", list, ...)
}
\arguments{
\item{x}{\code{QFeatures} file

@param assay_name
\code{Character}, define which assay needs to be extracted from QFeature input
e.g. "pos".}

\item{...}{Insert here parameter from \code{annotationNames} function from \code{MetNet}package.
(https://github.com/MetClassNet/MetNet)
\code{list} is the adjacency list, and \code{names} is a dataframe containing the feature names as rows,
mz values, RT values and annotations in columns}
}
\value{

}
\description{
The function \code{qfeat_annotation} uses the input from \code{QFeatures} and creates an
adjacency list adding manual annotations using \code{annotaionNames} from the \code{MetNet} package.
}
\details{

}
\examples{
####### To be added

}
\author{
Liesa Salzer,  \email{liesa.salzer@helmholtz-muenchen.de}
}
