% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qfeat.R
\name{qfeat_homol}
\alias{qfeat_homol}
\title{Homologue series extraction from QFeatures input}
\usage{
qfeat_homol(x, assay_name, ...)
}
\arguments{
\item{x}{\code{QFeatures} file}

\item{assay_name}{\code{Character}, define which assay needs to be extracted from QFeature input
e.g. "pos".}

\item{...}{Insert here parameter from \code{homol.search} function from \code{nontarget}package.
(https://cran.r-project.org/web/packages/nontarget/index.html)
Define parameters needed at \code{homol.search} as \code{elements}, \code{use_C},
\code{minmx}, \code{maxmz}, \code{minrt}, \code{maxrt}, \code{ppm}, \code{mztol}, \code{rttol}, \code{minlenght},
\code{mzfilter}, \code{spar}, \code{R2}.
For further information see Help page of \code{?homol.search}.}
}
\description{
The function \code{qfeat_homol} uses the input from \code{QFeatures} and uses dynamic
programming algorithms from \code{homol.search} from the \code{nontarget} package
(https://cran.r-project.org/web/packages/nontarget/index.html) for
unsupervides detection of homologue series in LC-(HR)MS data.
}
\details{
\code{qfeat_structural} extracts required information from a \code{QFeatures} input
and builds \code{list} containing results from \code{homol.search} function of
\code{nontarget} package.

@seealso \code{nontarget}, \code{homol.search}
https://cran.r-project.org/web/packages/nontarget/index.html

@return
\code{list} of type homol with 6 entries. First entry contains dataframe
with peaks (mass,intensity,rt,peak ID) and their homologue
series relations (to ID,m/z increment,RT increment) within different
homologue series (HS IDs,series level). Last column HS cluster states
HS clusters into which a peak was assigned via its HS. Second entry
contains the parameters used.
The third list entry is a Dataframe listing all peaks (peak IDs) per
homologue series (HS IDs), the underlying mean m/z & RT increments
(m/z increments, RT increments) and the minimum and maximum RT changes
between individual peaks of the series. The 4th entry stores used
m/z restrictions. List of peak IDs per level in the individual series
is stores in the 5th entry. The 6th entry contains List with
superjacent HS IDs per group - for setdeb=c(3,...).
}
\examples{
####### To be added

}
\author{
Liesa Salzer,  \email{liesa.salzer@helmholtz-muenchen.de}
}
