% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qfeat.R
\name{qfeat_statistical}
\alias{qfeat_statistical}
\title{Create a list of correlation matrices from QFeatures input}
\usage{
qfeat_statistical(x, model, p = F, ...)
}
\arguments{
\item{x}{\code{QFeatures} file}

\item{model}{\code{character} vector containing the methods that will be used
(\code{"lasso"}, \code{"randomForest"}, \code{"clr"}, \code{"aracne"}, \code{"pearson"},
\code{"pearson_partial"}, \code{"pearson_semipartial"}, \code{"spearman"},
\code{"spearman_partial"}, \code{"spearman_semipartial"}, \code{"bayes"})
\code{data.frame}, containing the columns \code{"group"}
(https://github.com/MetClassNet/MetNet).}

\item{p}{\code{logical}, by default is set to FALSE.}
}
\description{
The function \code{qfeat_statistical} uses the input from \code{QFeatures} and creates an
adjacency matrix based on statistical methods using the \code{MetNet} package.
The function includes functionality to calculate adjacency matrices based on
LASSO (L1 norm)-regression, random forests, context likelihood of
relatedness (CLR), the algorithm for the reconstruction of accurate
cellular networks (ARACNE), Pearson correlation (also partial and
semipartial), Spearman correlation (also partial and semipartial)
and score-based structure learning (Bayes). The function returns a
list of adjacency matrices that are defined by \code{model}.
Additionally, for pearson and/or spearman correlation also the negative
correlation values and the corresponding p-Value is calculated and listed,
when \code{p} is set to TRUE.
}
\details{
\code{qfeat_statistical} extracts required information from a \code{QFeatures} input
and builds data.frames containing intensity informations of all samples.
Then the function \code{statistical} from the \code{MetNet} package is applied
to calculate adjacency matrices based on
LASSO (L1 norm)-regression, random forests, context likelihood of
relatedness (CLR), the algorithm for the reconstruction of accurate
cellular networks (ARACNE), Pearson correlation (also partial and
semipartial), Spearman correlation (also partial and semipartial)
and Constraint-based structure learning (Bayes).
The default of \code{p} is FALSE. Then all types can be selected in \code{model}.
The positive correlation value will be displayed in the correlation matrix.
If \code{p} is set to TRUE, only "pearson" and/or "spearman" correlation may be
selected. As output positive and negative correlation values will be displayed.
Morover their corresponding p-values will be added to the list.

@return
\code{list} containing the respective adjacency matrices specified by
\code{model}.  It p\verb{is TRUE, also the corresponding p-values of Spearman and/or Pearson Correlation will be added to the}list`.
}
\examples{
####### To be added

}
\author{
Liesa Salzer,  \email{liesa.salzer@helmholtz-muenchen.de}
}
