% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qfeat.R
\name{qfeat_structural}
\alias{qfeat_structural}
\title{Create m/z difference adjacency matrix from QFeatures input}
\usage{
qfeat_structural(x, assay_name, ...)
}
\arguments{
\item{x}{\code{QFeatures} file

@param assay_name
\code{Character}, define which assay needs to be extracted from QFeature input
e.g. "pos".}

\item{...}{Insert here parameter from \code{structural} function from \code{MetNet}package.
(https://github.com/MetClassNet/MetNet)
\code{transformation} is a  \code{data.frame}, containing the columns \code{"group"},
and \code{"mass"} that will be used for detection of transformation of
(functional) groups
Parameter \code{ppm}is \code{numeric}, mass accuracy of m/z features in
parts per million (ppm)}
}
\description{
The function \code{qfeat_structural} uses the input from \code{QFeatures} and creates an
adjacency matrix based on m/z (molecular weight) difference from the \code{MetNet} package.
}
\details{
\code{qfeat_structural} extracts required information from a \code{QFeatures} input
and builds data.frames containing intensity informations, m/z values,
RT values, and annotations of all samples.
Then the function \code{structural} from the \code{MetNet} package is applied
to generate a m/z difference adjacency matrix.

@return
\code{list} containing three matrices. The first entry stores the \code{numeric}
\code{matrix} with edges inferred from mass differences. The second entry
stores the \code{character} \code{matrix} with the type (corresponding to the
\code{"group"} column in \code{transformation}) is stored. The third matrix stores
another \code{character} \code{matrix} with the corresponding mass difference (\code{"mass"}
column in \code{transformation}).
}
\examples{
####### To be added

}
\author{
Liesa Salzer,  \email{liesa.salzer@helmholtz-muenchen.de}
}
