% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms2similarity.R
\name{spec_adjacency_list}
\alias{spec_adjacency_list}
\title{Create an adjacency list from spectral similarity matrix}
\usage{
spec_adjacency_list(x)
}
\arguments{
\item{x}{\code{data.frame} adjacency matrix}
}
\value{
\code{data.frame} adjacency list
}
\description{
The function \code{spec_adjacency_list} creates an adjacency lists from
spectral similarity matrix and is comparable to MetNet function adjacency_list()
}
\examples{
 ######  example to be added!


}
\author{
Liesa Salzer, \email{liesa.salzer@helmholtz-muenchen.de}
}
