% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiLayerFunctions.R
\name{writeMultiLayer}
\alias{writeMultiLayer}
\title{Save and visualize multi-layer network}
\usage{
writeMultiLayer(inputData, multiLayer, visualize = FALSE)
}
\arguments{
\item{inputData}{\code{list}, list returned by the \code{loadInputData} function}

\item{multiLayer}{\code{list}, list containing the multi-layer network as returned by the
\code{makeMultiLayer} function}

\item{visualize}{\code{boolean}, if == TRUE, the multi-layer layer will be visualized in Cytoscape
(NOTE. Cytoscape needs to be open). The default value is FALSE}
}
\value{
Nothing, but it generates files with the list of nodes, edges, and the
Cytoscape visualization (if visualize == TRUE)
}
\description{
Function to save the list of edges and nodes from the multi-layer network
to visualize it in Cytoscape
}
\examples{
# See the MultiLayerNetwork vignette

}
\author{
Elva Novoa, \email{elva-maria.novoa-del-toro@inrae.fr}
}
