% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapMetToGSMN.R
\name{mapMetToGSMN}
\alias{mapMetToGSMN}
\title{Map the experimental nodes to GSMN nodes}
\usage{
mapMetToGSMN(
  inputData,
  method = "metabolomics2network",
  resFile = "Res_Met2Net_MappedMet.txt",
  ...
)
}
\arguments{
\item{inputData}{\code{list}, list returned by the \code{loadInputData} function}

\item{method}{one of Metabolomics2Network, molecularFormula, accurateMass, class_chebi or class_chemont}

\item{...}{additional parameter depending on the chosen function}
}
\value{
Data frame with the mappings and ontology-based distances.
}
\description{
Function to map the identified experimental nodes to the
corresponding GSMN nodes.
}
\author{
Elva Novoa, Sarah Scharfenberg
}
