% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qfeat.R
\name{qfeat_homol}
\alias{qfeat_homol}
\title{Homologue series extraction from QFeatures input}
\usage{
qfeat_homol(x, assay_name = "features", plot = FALSE, ...)
}
\arguments{
\item{x}{\code{QFeatures} file}

\item{assay_name}{\code{Character}, define which assay needs to be extracted from QFeature input
e.g. "pos".}

\item{...}{Insert here parameter from \code{homol.search} function from \code{nontarget}package.
(https://cran.r-project.org/web/packages/nontarget/index.html)
Define parameters needed at \code{homol.search} as \code{elements}, \code{use_C},
\code{minmx}, \code{maxmz}, \code{minrt}, \code{maxrt}, \code{ppm}, \code{mztol}, \code{rttol}, \code{minlenght},
\code{mzfilter}, \code{spar}, \code{R2}.
For further information see Help page of \code{?homol.search}.}
}
\description{
The function \code{qfeat_homol} uses the input from \code{QFeatures} and uses dynamic
programming algorithms from \code{homol.search} from the \code{nontarget} package
(https://cran.r-project.org/web/packages/nontarget/index.html) for
unsupervides detection of homologue series in LC-(HR)MS data.
}
\details{
\code{qfeat_structural} extracts required information from a \code{QFeatures} input
and builds \code{list} containing results from \code{homol.search} function of
\code{nontarget} package.

@seealso \code{nontarget}, \code{homol.search}
https://cran.r-project.org/web/packages/nontarget/index.html

@return
\code{data.frame} edgelist of homologues series with columns "cluster_ID",
"cluster_ID_to", "HSIDs", "mzincrement", "RTincrement".
}
\examples{
####### To be added

}
\author{
Liesa Salzer,  \email{liesa.salzer@helmholtz-muenchen.de}
}
