% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms2similarity.R
\name{spec_adjacency_list}
\alias{spec_adjacency_list}
\alias{spec_df}
\title{Create a \code{data.frame} from spectral similarity matrix}
\usage{
spec_df(x)
}
\arguments{
\item{x}{\code{data.frame} adjacency matrix}
}
\value{
\code{data.frame} adjacency list
}
\description{
The function \code{spec_df} creates a \code{data.frame} from
spectral similarity matrix
}
\examples{
 ######  example to be added!


}
\author{
Liesa Salzer, \email{liesa.salzer@helmholtz-muenchen.de}
}
